Update tests for CreateWallet
